requirements.txt
# Core data analysis
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.11.0

# Statistical analysis
statsmodels>=0.13.0
scikit-learn>=1.1.0
pingouin>=0.5.3  # Advanced statistical tests

# Data handling
openpyxl>=3.0.10  # Excel file support
xlsxwriter>=3.0.0  # Excel writing

# Utilities
python-dateutil>=2.8.0
pytz>=2022.1
tqdm>=4.64.0  # Progress bars

# Development tools
jupyter>=1.0.0
jupyterlab>=3.4.0
ipykernel>=6.15.0
black>=22.0.0  # Code formatting
pylint>=2.15.0  # Code linting
pytest>=7.1.0  # Testing

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Optional but recommended
tabulate>=0.9.0  # Pretty table printing
click>=8.1.0  # CLI interface
python-dotenv>=0.21.0  # Environment variables